Fix most issues, but 2x spike remains at end
I think I managed to set the guard cells to zero somehow. There is a problem with the gradient which freaks out near the end and if you use simpson's rule you catch some of that and cause a spike. I really want to set yend equal to yend-1 but idk how to do that...