Feature: Fcx_exc and Frec_sk
Fcx_exc is the charge exchange friction contribution from the 30 excited states of SOLKIT. Higher states drop off in population by n^2 but increase in cross-section by n^4 so it's really important. Frec_sk is the mom sink due to recombination in SOLKiT. They have a CRM and track excited states, which are really easy to ionise/recombine so they do this a lot. Because we have diffusive neutrals, plasma momentum is lost in recombination but not gained in ionisation.